gather: move EsExporter to own file
